reset pos

+++ Assets/Scripts/GoapHeroActions/GoapHeroWipeBloodAction.cs
@@ -34,7 +34,7 @@ namespace Assets.Scripts.GoapHeroActions
         }
         public override bool perform(GameObject agent)
         {
-            if (GetActiveNpcAttributesComponentsInRange(gameObject, PoseThreshold) < 1)
+            if (GetActiveNpcAttributesComponentsInRange(gameObject, PoseThreshold) < 1 && InResetRange())
             {
                 HeroScript.WipeBlood();
                 NpcIsDestroyedReset = true;
